refactor(study): repository 타입 수정 Integer -> Long DP-266

+++ src/main/java/goorm/ddok/study/repository/StudyApplicationRepository.java
@@ -3,5 +3,5 @@ package goorm.ddok.study.repository;
 import goorm.ddok.study.domain.StudyApplication;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StudyApplicationRepository extends JpaRepository<StudyApplication, Integer> {
+public interface StudyApplicationRepository extends JpaRepository<StudyApplication, Long> {
 }
